refactoring BaseStatusPage: - added locators  header, description - added methods getNameText(), getDescriptionText()
removed duplicate code from Page classes
