+++ src/test/java/model/multibranch_pipeline/MultibranchPipelineStatusPage.java
@@ -11,9 +11,6 @@ public class MultibranchPipelineStatusPage extends BaseStatusPage {
     @FindBy(linkText = "Delete Multibranch Pipeline")
     private WebElement deleteLeftSideMenu;
 
-    @FindBy(xpath = "//h1")
-    private WebElement multibranchPipelineName;
-
     @FindBy(linkText = "Rename")
     private WebElement renameLeftMenu;
 
@@ -25,10 +22,6 @@ public class MultibranchPipelineStatusPage extends BaseStatusPage {
         deleteLeftSideMenu.click();
 
         return new DeleteMultibranchPipelinePage(getDriver());
-    }
-
-    public String getMultibranchPipelineName() {
-        return multibranchPipelineName.getText();
     }
 
     public RenameItemPage clickRenameLeftMenu() {
