+++ src/test/java/model/multiconfiguration/MultiConfigurationProjectStatusPage.java
@@ -23,9 +23,6 @@ public class MultiConfigurationProjectStatusPage extends BaseStatusPage {
     @FindBy(xpath = "//button[contains(text(),'Save')]")
     private WebElement saveDescriptionButton;
 
-    @FindBy(xpath = "//div[@id='description']/div[1]")
-    private WebElement fieldDescription;
-
     @FindBy(xpath = "//span[text()='Delete Multi-configuration project']")
     private WebElement deleteOption;
 
@@ -83,11 +80,6 @@ public class MultiConfigurationProjectStatusPage extends BaseStatusPage {
         saveDescriptionButton.click();
 
         return this;
-    }
-
-    public String getDescriptionText() {
-
-        return fieldDescription.getText();
     }
 
     public String getNameMultiConfigProject(String name) {
