+++ src/test/java/tests/FolderOneTest.java
@@ -41,7 +41,7 @@ public class FolderOneTest extends BaseTest {
                 .selectFolderAndClickOk()
                 .clickSaveBtn(FolderStatusPage.class);
 
-        Assert.assertTrue(folderStatusPage.getHeaderText().contains(RANDOM_NAME_2));
+        Assert.assertTrue(folderStatusPage.getNameText().contains(RANDOM_NAME_2));
         Assert.assertTrue(folderStatusPage.getTopMenueLinkText().contains(RANDOM_NAME_2));
         Assert.assertTrue(folderStatusPage.getTopMenueLinkText().contains(RANDOM_NAME_1));
     }
@@ -111,7 +111,7 @@ public class FolderOneTest extends BaseTest {
         FolderStatusPage statusPage = new HomePage(getDriver())
                 .clickFolder(RANDOM_NAME_2);
 
-        Assert.assertTrue(statusPage.getHeaderText().contains(RANDOM_NAME_2));
+        Assert.assertTrue(statusPage.getNameText().contains(RANDOM_NAME_2));
         Assert.assertTrue(statusPage.getTopMenueLinkText().contains(RANDOM_NAME_1));
         Assert.assertTrue(statusPage.getTopMenueLinkText().contains(RANDOM_NAME_2));
     }
@@ -170,7 +170,7 @@ public class FolderOneTest extends BaseTest {
 
         Assert.assertTrue(actualResult.getTopMenueLinkText().contains(RANDOM_NAME_1));
         Assert.assertTrue(actualResult.getTopMenueLinkText().contains(RANDOM_MULTIBRANCH_PIPELINE_NAME));
-        Assert.assertTrue(actualResult.getHeaderText().contains(RANDOM_MULTIBRANCH_PIPELINE_NAME));
+        Assert.assertTrue(actualResult.getNameText().contains(RANDOM_MULTIBRANCH_PIPELINE_NAME));
     }
 
     @Test
@@ -203,7 +203,7 @@ public class FolderOneTest extends BaseTest {
                 .setDisplayName("NewDisplayName")
                 .clickSaveBtn(FolderStatusPage.class);
 
-        Assert.assertTrue(statusPage.getHeaderText().contains("NewDisplayName"));
+        Assert.assertTrue(statusPage.getNameText().contains("NewDisplayName"));
         Assert.assertTrue(statusPage.getTopMenueLinkText().contains(RANDOM_NAME_1));
         Assert.assertTrue(statusPage.getTopMenueLinkText().contains("NewDisplayName"));
     }
@@ -237,7 +237,7 @@ public class FolderOneTest extends BaseTest {
                 .clickDashboard()
                 .clickFolder(RANDOM_NAME_1);
 
-        Assert.assertEquals(folderStatusPage.getHeaderFolderText(), RANDOM_NAME_1);
+        Assert.assertEquals(folderStatusPage.getNameText(), RANDOM_NAME_1);
         Assert.assertTrue(folderStatusPage.getJobList().size() != 0);
         Assert.assertTrue(folderStatusPage.getJobList().contains(RANDOM_MULTIBRANCH_PIPELINE_NAME));
     }
@@ -276,7 +276,7 @@ public class FolderOneTest extends BaseTest {
                 .clickDeleteMultibranchPipeline()
                 .clickSubmitButton();
 
-        Assert.assertEquals(folderStatusPage.getHeaderFolderText(), RANDOM_NAME_1);
+        Assert.assertEquals(folderStatusPage.getNameText(), RANDOM_NAME_1);
         Assert.assertNotNull(folderStatusPage.getEmptyStateBlock());
     }
 }
