+++ src/test/java/tests/FolderTest.java
@@ -80,7 +80,7 @@ public class FolderTest extends BaseTest {
                 .selectFolderAndClickOk()
                 .setDescription(addDescription)
                 .clickSaveBtn(FolderStatusPage.class)
-                .getTextDescription();
+                .getDescriptionText();
 
         Assert.assertEquals(textDescription, addDescription);
     }
@@ -107,7 +107,7 @@ public class FolderTest extends BaseTest {
                 .clickJob(DISPLAY_RANDOM_NAME)
                 .clickDeleteFolder()
                 .clickSubmitButton()
-                .getHeaderText();
+                .getNameText();
 
         Assert.assertEquals(pageHeaderText, "Welcome to Jenkins!");
     }
@@ -230,7 +230,7 @@ public class FolderTest extends BaseTest {
                 .clickSubmitButton()
                 .clickDashboard()
                 .clickFolder(FOLDER_RANDOM_NAME_1)
-                .getTextDescriptionOnPage();
+                .getDescriptionText();
 
         Assert.assertEquals(textDescription, folderDescription);
     }
