+++ src/test/java/tests/MultibranchPipelineTest.java
@@ -101,7 +101,7 @@ public class MultibranchPipelineTest extends BaseTest {
                 .clickRenameDropDownMenu()
                 .clearFieldAndInputNewName(MULTIBRANCH_PIPELINE_NAME_RENAMED)
                 .clickRenameButtonMBPipeline()
-                .getMultibranchPipelineName();
+                .getNameText();
 
         Assert.assertTrue(actualMultibranchPipeline.contains(MULTIBRANCH_PIPELINE_NAME_RENAMED));
     }
@@ -113,7 +113,7 @@ public class MultibranchPipelineTest extends BaseTest {
                 .clickRenameLeftMenu()
                 .clearFieldAndInputNewName(MULTIBRANCH_PIPELINE_NAME_RENAMED)
                 .clickRenameButtonMBPipeline()
-                .getMultibranchPipelineName();
+                .getNameText();
 
         Assert.assertTrue(actualMultibranchPipeline.contains(MULTIBRANCH_PIPELINE_NAME_RENAMED));
