+++ src/test/java/tests/PipelineTest.java
@@ -66,7 +66,7 @@ public class PipelineTest extends BaseTest {
                 .clearFieldAndInputNewName(PIPELINE_NAME + RENAME_SUFFIX)
                 .clickSubmitButton();
 
-        Assert.assertEquals(new PipelineStatusPage(getDriver()).getPipelineTitle(), "Pipeline " + PIPELINE_NAME + RENAME_SUFFIX);
+        Assert.assertEquals(new PipelineStatusPage(getDriver()).getNameText(), "Pipeline " + PIPELINE_NAME + RENAME_SUFFIX);
     }
 
     @Test
@@ -168,7 +168,7 @@ public class PipelineTest extends BaseTest {
                 .clickSaveButton()
                 .clickDashboard().clickPipelineJob(PIPELINE_NAME);
 
-        Assert.assertEquals(pipelineProjectPage.getDescription(), PIPELINE_NAME + "edit description");
+        Assert.assertEquals(pipelineProjectPage.getDescriptionText(), PIPELINE_NAME + "edit description");
     }
 
     @Test(dependsOnMethods = "testEnablePipelineProject")
@@ -215,7 +215,7 @@ public class PipelineTest extends BaseTest {
                 .setItemName(PIPELINE_NAME)
                 .selectPipelineAndClickOk()
                 .clickSaveBtn(PipelineStatusPage.class)
-                .getPipelineTitle();
+                .getNameText();
 
         Assert.assertEquals(actualPipelineName, "Pipeline " + PIPELINE_NAME);
     }
@@ -289,7 +289,7 @@ public class PipelineTest extends BaseTest {
                 .editDescription(PIPELINE_DESCRIPTION)
                 .clickSaveButton();
 
-        Assert.assertEquals(pipelineProjectPage.getDescription(), PIPELINE_DESCRIPTION);
+        Assert.assertEquals(pipelineProjectPage.getDescriptionText(), PIPELINE_DESCRIPTION);
     }
 
     @Test
@@ -300,7 +300,7 @@ public class PipelineTest extends BaseTest {
                 .selectPipelineAndClickOk()
                 .setDescriptionField(PIPELINE_DESCRIPTION)
                 .clickSaveBtn(PipelineStatusPage.class)
-                .getDescription();
+                .getDescriptionText();
 
         Assert.assertEquals(actualPipelineDescription, PIPELINE_DESCRIPTION);
     }
@@ -313,7 +313,7 @@ public class PipelineTest extends BaseTest {
                 .clickPipelineProjectName()
                 .editDescription(ITEM_NEW_DESCRIPTION)
                 .clickSaveButton()
-                .getDescription();
+                .getDescriptionText();
 
         Assert.assertEquals(actualDescription, ITEM_NEW_DESCRIPTION);
     }
@@ -328,7 +328,7 @@ public class PipelineTest extends BaseTest {
                 .clickSaveBtn(PipelineStatusPage.class)
                 .getPipelineName();
 
-        String actualDescription = new PipelineStatusPage(getDriver()).getDescription();
+        String actualDescription = new PipelineStatusPage(getDriver()).getDescriptionText();
 
         Assert.assertEquals(actualJobName, ITEM_NAME);
         Assert.assertEquals(actualDescription, ITEM_NEW_DESCRIPTION);
@@ -397,7 +397,7 @@ public class PipelineTest extends BaseTest {
                 .clickConfigure()
                 .setDescriptionField(ITEM_NEW_DESCRIPTION)
                 .clickSaveBtn(PipelineStatusPage.class)
-                .getDescription();
+                .getDescriptionText();
 
         Assert.assertEquals(actualDescription, ITEM_NEW_DESCRIPTION);
     }
