+++ src/test/java/tests/ProjectsInFolderTest.java
@@ -30,7 +30,7 @@ public class ProjectsInFolderTest extends BaseTest {
                 .setItemName(RANDOM_NAME)
                 .selectOrgFolderAndClickOk()
                 .clickSaveBtn(OrgFolderStatusPage.class)
-                .getDisplayName();
+                .getNameText();
 
         Assert.assertEquals(actualOrganizationFolderDisplayName, RANDOM_NAME);
     }
